OP-14 OP-16: I2C Improvements 1) Added nack counter monitoring 2) Made timeout for getting semaphore in I2C user space code use the one from driver and record timeouts.  This does not influence timeouts in the non FreeRTOS case 3) Remove case block from the error handler so that all bus errors reset the i2c interface
git-svn-id: svn://svn.openpilot.org/OpenPilot/trunk@2469 ebee16cc-31ac-478f-84a7-5cbb03baadba

+++ flight/PiOS/inc/pios_i2c.h
@@ -69,7 +69,7 @@ extern int32_t PIOS_I2C_Init(void);
 extern bool PIOS_I2C_Transfer(uint8_t i2c, const struct pios_i2c_txn txn_list[], uint32_t num_txns);
 extern void PIOS_I2C_EV_IRQ_Handler(uint8_t i2c);
 extern void PIOS_I2C_ER_IRQ_Handler(uint8_t i2c);
-extern void PIOS_I2C_GetDiagnostics(struct pios_i2c_fault_history * data, uint16_t * error_counts);
+extern void PIOS_I2C_GetDiagnostics(struct pios_i2c_fault_history * data, uint8_t * error_counts);
 
 #endif /* PIOS_I2C_H */
